+++ packages/yoroi-extension/app/api/ada/index.js
@@ -42,12 +42,12 @@ import {
   asGetAllUtxos,
   asGetPublicKey,
   asGetSigningKey,
+  asGetStakingKey,
   asHasLevels,
   asHasUtxoChains,
-  asGetStakingKey,
 } from './lib/storage/models/PublicDeriver/traits';
 import { ConceptualWallet } from './lib/storage/models/ConceptualWallet/index';
-import { WalletTypeOption, type IHasLevels } from './lib/storage/models/ConceptualWallet/interfaces';
+import { type IHasLevels, WalletTypeOption } from './lib/storage/models/ConceptualWallet/interfaces';
 import type {
   Address,
   Addressing,
@@ -56,8 +56,6 @@ import type {
   IDisplayCutoff,
   IGetAllUtxoAddressesResponse,
   IGetAllUtxosResponse,
-  IGetSigningKey,
-  IGetStakingKey,
   IHasUtxoChains,
   IHasUtxoChainsRequest,
   IPublicDeriver,
@@ -2571,37 +2569,6 @@ function getDifferenceAfterTx(
   }
 
   return sumOutForKey.joinSubtractCopy(sumInForKey);
-}
-
-export async function genOwnStakingKey(request: {|
-  publicDeriver: IPublicDeriver<ConceptualWallet & IHasLevels> & IGetSigningKey & IGetStakingKey,
-  password: string,
-|}): Promise<RustModule.WalletV4.PrivateKey> {
-  try {
-    const signingKeyFromStorage = await request.publicDeriver.getSigningKey();
-    const stakingAddr = await request.publicDeriver.getStakingKey();
-    const normalizedKey = await request.publicDeriver.normalizeKey({
-      ...signingKeyFromStorage,
-      password: request.password,
-    });
-    const normalizedSigningKey = RustModule.WalletV4.Bip32PrivateKey.from_hex(normalizedKey.prvKeyHex);
-    const normalizedStakingKey = derivePrivateByAddressing({
-      addressing: stakingAddr.addressing,
-      startingFrom: {
-        key: normalizedSigningKey,
-        level: request.publicDeriver.getParent().getPublicDeriverLevel(),
-      },
-    }).to_raw_key();
-
-    return normalizedStakingKey;
-  } catch (error) {
-    Logger.error(`${nameof(genOwnStakingKey)} error: ` + stringifyError(error));
-    if (error instanceof WrongPassphraseError) {
-      throw new IncorrectWalletPasswordError();
-    }
-    if (error instanceof LocalizableError) throw error;
-    throw new GenericApiError();
-  }
 }
 
 export { cip8Sign } from './lib/cardanoCrypto/utils';
